requirements.txt
flask==3.0.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.32
python-dotenv==1.0.1
requests==2.32.3
pymysql==1.1.1

